pyproject.toml
[project]
name = "withings-mcp-server"
version = "0.1.0"
description = "MCP Server for Withings API integration"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/withings_mcp_server"]

[project.scripts]
withings-mcp-server = "withings_mcp_server.server:main"
